TopicModelling: Fix progress bar issues
The progress percentage in window's title was not updated properly but the percentage under the widget in canvas worked ok. The problem was that we just called the widget's method from an other thread while one should not do this (tnx @ales-erjavec). This commit fixes this so that progress is properly called through slots.