Corpus: Fix extend_attributes for dense matrices
Extending dense matrix with an other dense matrix does not work with scipy.sparse.hstack not do we want to transform this to sparse. To reproduce use Topic Modelling on Twitter data with Author field checked. This PR fixes extending by using numpy  or spicpy sparse hstach depending on matrices sparseness or denseness.