NYT: TimeVariable as class var
This fixed tests for Orange's versions above 3.3.8. Due to using different instaces for parsing, the instace in domain did not have set `have_time` nor `have_date` attribute which caused `repr_val` function to just return string representation of the float. To fix the problem, we now use the same instance in domain as for parsing.