StopwordsFilter: Don't crash when no NLTK data
If Orange was started without internet connection NLTK data was not downloaded. When the data is missing, importing anything from `orangecontrib.text.preprocess` crashed with a `LookupError` since `StopwordsFilter` assumed the data is alredy downloaded. Use empy list as a list of supported stopwords languages in `StopwordsFilter` when NLTK data is missing.