Corpus: Use given X even if empty
If X is given use it no matter if it contains data or not. This is required for applying a BoW classfifer on new data. Classifiers use `from_table` (which creates a new corpus) and if no classifier's tokens appear in the new data documents, then X is an empty sparse matrix of the correct shape. In this case, we currently replaced this empty X with a sparse matrix without columns which caused classifers to crash with shape mismatch (but only when there was no words overlap).